Refactor handler file
